Make fluent and fluent-syntax separate packages
