+++ .travis.yml
@@ -1,5 +1,6 @@
 sudo: false
 language: node_js
+script: make all
 node_js:
   - "6.0.0"
 notifications:
