+++ README.md
@@ -5,8 +5,7 @@ Project Fluent [![Build Status][travisimage]][travislink]
 [travislink]: https://travis-ci.org/projectfluent/fluent.js
 
 Fluent.js is a JavaScript implementation of Project Fluent, a localization
-framework designed to unleash the entire expressive power of natural language
-translations.
+framework designed to unleash the expressive power of the natural language.
 
 Project Fluent keeps simple things simple and makes complex things possible.
 The syntax used for describing translations is easy to read and understand.  At
@@ -14,55 +13,14 @@ the same time it allows, when necessary, to represent complex concepts from
 natural languages like gender, plurals, conjugations, and others.
 
 
-Installation
-------------
+Packages
+--------
 
-Fluent.js can be used both on the client-side and the server-side.  Install the
-`fluent` package from the npm registry:
+Fluent.js consists of a set of packages which have different use-cases and can
+be installed independently of each other.
 
-    npm install fluent
-
-Fluent.js also works as a standalone script (as the `Fluent` global) and with
-`RequireJS`.
-
-
-How to use Fluent.js
---------------------
-
-When you import the `fluent` module it will add the `Intl.MessageContext`
-constructor to the global `Intl` object, together with a few other polyfills.
-You can consult the complete list of the polyfills in [src/intl/polyfill.js]. 
-
-`Intl.MessageContext` provides the core functionality of formatting
-translations from FTL files.
-
-[src/intl/polyfill.js]: https://github.com/projectfluent/fluent.js/blob/master/src/intl/polyfill.js
-
-```javascript
-require('fluent');
-
-const ctx = new Intl.MessageContext('en-US');
-
-const errors = ctx.addMessages(`
-brand-name = Foo 3000
-welcome    = Welcome, { $name }, to { brand-name }!
-`);
-
-if (errors.length) {
-  // syntax errors are per-message and don't break the whole resource
-}
-
-const welcome = ctx.messages.get('welcome');
-
-ctx.format(welcome, { name: 'Anna' });
-// → 'Welcome, Anna, to Foo 3000!'
-```
-
-You can access other API methods by requiring the `fluent` module.  The
-currently exposed functions can be found in [src/index.js][].  Keep in mind
-that `fluent` is in its 0.x days and that the exposed API may change.
-
-[src/index.js]: https://github.com/projectfluent/fluent.js/blob/master/src/index.js
+  - [fluent](https://github.com/projectfluent/fluent.js/tree/master/fluent)
+  - [fluent-syntax](https://github.com/projectfluent/fluent.js/tree/master/fluent-syntax)
 
 
 Learn the FTL syntax
