+++ fluent-syntax/.gitignore
@@ -1,0 +1,1 @@
+fluent-syntax.js
