+++ fluent-syntax/.npmignore
@@ -1,4 +1,2 @@
-docs
 test
-tools
 makefile
