+++ fluent-syntax/src/ftlstream.js
