+++ fluent-syntax/src/index.js
@@ -1,0 +1,2 @@
+export * from './ast';
+export { parse } from './parser';
