+++ fluent-syntax/src/stream.js
