+++ fluent/src/plurals.js
