+++ fluent/src/resolver.js
