+++ fluent/test/compat.js
@@ -1,0 +1,5 @@
+'use strict';
+
+require('babel-register')({
+  plugins: ['transform-es2015-modules-commonjs']
+});
