+++ fluent/test/patterns_test.js
@@ -2,8 +2,8 @@
 
 import assert from 'assert';
 
-import { MessageContext } from '../../src/intl/context';
-import { ftl } from '../util';
+import { MessageContext } from '../src/context';
+import { ftl } from './util';
 
 describe('Patterns', function(){
   let ctx, args, errs;
