+++ package.json
@@ -1,35 +1,7 @@
 {
-  "name": "fluent",
-  "description": "Localization library for expressive translations.",
-  "version": "0.2.1",
-  "homepage": "http://projectfluent.io",
-  "author": "Mozilla <l10n-drivers@mozilla.org>",
-  "license": "Apache-2.0",
-  "contributors": [
-    {
-      "name": "Zibi Braniecki",
-      "email": "zbraniecki@mozilla.com"
-    },
-    {
-      "name": "Staś Małolepszy",
-      "email": "stas@mozilla.com"
-    }
-  ],
-  "scripts": {
-    "test": "make"
-  },
-  "directories": {
-    "lib": "./src"
-  },
-  "main": "./dist/fluent.js",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/projectfluent/fluent.js.git"
-  },
-  "keywords": [
-    "localization",
-    "l10n"
-  ],
+  "name": "fluent-javascript",
+  "version": "0.0.1",
+  "private": true,
   "devDependencies": {
     "babel-plugin-transform-es2015-modules-commonjs": "^6.7.7",
     "babel-register": "^6.7.2",
@@ -41,157 +13,5 @@
     "mocha": "^3.0.2",
     "prettyjson": "^1.1.3",
     "rollup": "^0.41.4"
-  },
-  "engine": {
-    "node": ">=6"
-  },
-  "eslintConfig": {
-    "parserOptions": {
-      "ecmaVersion": 6,
-      "sourceType": "module"
-    },
-    "env": {
-      "es6": true
-    },
-    "rules": {
-      "valid-typeof": 2,
-      "no-console": 1,
-      "no-dupe-args": 2,
-      "no-dupe-keys": 2,
-      "no-duplicate-case": 2,
-      "no-extra-semi": 1,
-      "no-func-assign": 2,
-      "func-call-spacing": [
-        2,
-        "never"
-      ],
-      "no-unreachable": 1,
-      "no-unexpected-multiline": 2,
-      "no-sparse-arrays": 2,
-      "block-scoped-var": 2,
-      "complexity": [
-        1,
-        18
-      ],
-      "dot-location": [
-        2,
-        "property"
-      ],
-      "no-implied-eval": 2,
-      "no-loop-func": 2,
-      "no-magic-numbers": [
-        1,
-        {
-          "ignore": [
-            -1,
-            0,
-            1,
-            2
-          ]
-        }
-      ],
-      "no-useless-call": 2,
-      "no-useless-concat": 2,
-      "no-delete-var": 2,
-      "no-shadow": 2,
-      "no-undef-init": 2,
-      "no-unused-vars": 1,
-      "consistent-return": 1,
-      "curly": [
-        2,
-        "multi-line"
-      ],
-      "eqeqeq": 2,
-      "no-extend-native": 2,
-      "no-global-assign": 2,
-      "no-extra-bind": 2,
-      "no-redeclare": 2,
-      "array-bracket-spacing": 2,
-      "brace-style": [
-        1,
-        "1tbs"
-      ],
-      "no-mixed-spaces-and-tabs": 2,
-      "no-tabs": 2,
-      "prefer-arrow-callback": 1,
-      "prefer-const": 2,
-      "prefer-template": 2,
-      "prefer-spread": 2,
-      "quotes": [
-        2,
-        "single",
-        {
-          "avoidEscape": true
-        }
-      ],
-      "no-var": 2,
-      "max-depth": [
-        2,
-        6
-      ],
-      "max-len": [
-        2,
-        80
-      ],
-      "no-nested-ternary": 2,
-      "no-unneeded-ternary": 2,
-      "strict": [
-        2,
-        "global"
-      ],
-      "indent": [
-        2,
-        2,
-        {
-          "SwitchCase": 1
-        }
-      ],
-      "no-useless-escape": 2,
-      "no-duplicate-imports": 2,
-      "no-unsafe-negation": 2,
-      "no-use-before-define": [
-        2,
-        {
-          "functions": false
-        }
-      ],
-      "space-infix-ops": 2,
-      "no-constant-condition": [
-        1,
-        {
-          "checkLoops": false
-        }
-      ],
-      "no-empty": 1,
-      "use-isnan": 2,
-      "dot-notation": 2,
-      "no-caller": 2,
-      "no-else-return": 2,
-      "no-eval": 2,
-      "no-implicit-globals": 2,
-      "no-iterator": 2,
-      "no-multi-spaces": 2,
-      "no-multi-str": 2,
-      "no-octal": 2,
-      "no-proto": 2,
-      "no-sequences": 2,
-      "block-spacing": 2,
-      "eol-last": 2,
-      "key-spacing": 2,
-      "no-trailing-spaces": 2,
-      "prefer-rest-params": 2,
-      "no-cond-assign": 2,
-      "keyword-spacing": 2,
-      "arrow-parens": [
-        2,
-        "as-needed"
-      ],
-      "no-useless-return": 2,
-      "semi": 2,
-      "spaced-comment": [
-        2,
-        "always"
-      ]
-    }
   }
 }
